✅ Add multicast reciever test
